facial-expression-app/requirements.txt
streamlit==1.34.0
opencv-python-headless==4.9.0.80
tensorflow==2.15.0
numpy==1.24.4
Pillow==10.2.0